Move constants and add namespace
