+++ features/facebookImport/src/constants.js
@@ -1,0 +1,5 @@
+export const INITIAL_HISTORY_STATE = {
+    scrollingProgress: 0,
+};
+
+export const FBIMPORT_NAMESPACE = "http://polypoly.coop/schema/fbImport/";
