+++ features/facebookImport/src/facebookImporter.jsx
@@ -35,7 +35,7 @@ import manifestData from "./static/manifest.json";
 window.manifestData = manifestData;
 
 import i18n from "./i18n.js";
-import { INITIAL_HISTORY_STATE } from "../constants/constants.js";
+import { INITIAL_HISTORY_STATE } from "./constants.js";
 import {
     FileLoaderContext,
     FileLoaderProvider,
